Css changes

+++ client/src/components/CommentCard.vue
@@ -3,6 +3,7 @@ import { computed } from 'vue';
 import { Comment } from '../models/Comment.js';
 import { AppState } from '../AppState.js';
 import { commentsService } from '../services/CommentsService.js';
+import Pop from '../utils/Pop.js';
 
 
 const props = defineProps({ comment: { type: Comment, required: true } })
@@ -11,6 +12,8 @@ const account = computed(() => AppState.account)
 
 async function deleteComment(commentId) {
     try {
+        const wantToDestroy = await Pop.confirm("Are you sure you want to delete this comment?")
+        if (!wantToDestroy) return
         await commentsService.deleteComment(commentId)
     } catch (error) {
         console.error(error)
@@ -25,12 +28,13 @@ async function deleteComment(commentId) {
     <section class="row bg-light rounded m-1">
 
         <div class="col-12 col-sm-2">
-            <div class="m-3">
-                <img class="img-fluid" :src="comment.creator?.picture" :alt="comment.creator.picture">
+            <div class="m-2">
+                <img class="img-fluid profile-img-circle" :src="comment.creator?.picture"
+                    :alt="comment.creator.picture">
             </div>
         </div>
         <div class="col-12 col-sm-10">
-            <div class="mt-2">
+            <div class="mt-2 text-dark">
                 <h5>
                     {{ comment.creator.name }}
                 </h5>
